Added preferences.js module
A simple class that manages preferences stored in localStorage. It has to be instantiated with an object with the preferences that are going to be handled and with their default values. The preference's key will be set as properties of the instance (setters and getters) for easy access and modification.